+++ examples/with-redux-observable/pages/_app.js
@@ -1,5 +1,5 @@
 import { Provider } from 'react-redux'
-import { useStore } from '../redux/store'
+import { useStore } from '../store/store'
 
 export default function App({ Component, pageProps }) {
   const store = useStore(pageProps.initialReduxState)
