+++ examples/with-redux-observable/store/actionTypes.js
@@ -1,5 +1,5 @@
-export const FETCH_CHARACTER = 'FETCH_CHARACTER'
-export const FETCH_CHARACTER_SUCCESS = 'FETCH_CHARACTER_SUCCESS'
-export const FETCH_CHARACTER_FAILURE = 'FETCH_CHARACTER_FAILURE'
-export const START_FETCHING_CHARACTERS = 'START_FETCHING_CHARACTERS'
-export const STOP_FETCHING_CHARACTERS = 'STOP_FETCHING_CHARACTERS'
+export const FETCH_USER = 'FETCH_USER'
+export const FETCH_USER_SUCCESS = 'FETCH_USER_SUCCESS'
+export const FETCH_USER_FAILURE = 'FETCH_USER_FAILURE'
+export const START_FETCHING_USERS = 'START_FETCHING_USERS'
+export const STOP_FETCHING_USERS = 'STOP_FETCHING_USERS'
